+++ Changelog.md
@@ -9,6 +9,7 @@
 ### Fixed
 - Fixed delete authenticator API
 - Fixed deletion of linked SSI service while deleting IdService
+- Fixed stale tenant permissions in generated service access tokens after role updates.
 ## [3.11.13] - 2026-05-14
 ### Added 
 - Update package.json to bump version to 3.11.13.